[WIP] Change timestamps type to OffsetDateTime in Job/Step Execution
This is an attempt to change the data type of startTime, endTime,
createTime and lastUpdated in JobExecution and StepExecution to
OffsetDateTime, which corresponds to the SQL type TIMESTAMP WITH TIME ZONE.

However, the type `TIMESTAMP WITH TIME ZONE` does not seem to be supported
by some databases (or their drivers). Here is a quick list of issues related
to this change:

### MySQL:

```
Unsupported SQL type: TIMESTAMP_WITH_TIMEZONE
```

At `JdbcJobExecutionDao` (L173 as of 2b258d1b88661c581c6f79294e9c1f96ed69a159)

```
getJdbcTemplate().update(getQuery(SAVE_JOB_EXECUTION), parameters,
   new int[] { Types.BIGINT, Types.BIGINT, Types.TIMESTAMP_WITH_TIMEZONE,
   Types.TIMESTAMP_WITH_TIMEZONE, Types.VARCHAR, Types.VARCHAR, Types.VARCHAR,
   Types.INTEGER, Types.TIMESTAMP_WITH_TIMEZONE, Types.TIMESTAMP_WITH_TIMEZONE });
```

### DB2:

```
[jcc][1091][10417][4.31.10] Invalid data conversion: Parameter instance 2022-10-11T15:21:26.905406+02:00 is invalid for the requested conversion. ERRORCODE=-4461, SQLSTATE=42815
```

### Derby:

```
ERROR 22005: An attempt was made to get a data value of type 'TIMESTAMP' from a data value of type '2014'.
```

---

Adding custom code for specific DB providers in various DAOs is not an option
for the time being.

For all these reasons, it was decided to defer the implementation of this
feature to future release.

Issue #1014