Change visibility of properties to protected in JdbcPagingItemReaderBuilder
There are cases, when you want to extend the JdbcPagingItemReader
by adding some custom fields and behaviours, but want to stay in
the normal way of creating this class via a Builder class.

If you want to do it, you currently have to reimplement the whole
class just for adding some fields because there is no non-ugly way
to access the fields. This commit makes those properties protected,
so one can override the default build() method behaviour in the
derived class and still have access to the properties.

Issue #4331