Changed visibility of propertys from the JdbcPagingItemReaderBuilder.java from private to protected
There are cases, when you want to extend the JdbcPagingItemReader by adding some custom fields and behaviours, but want to stay in the normal way of creating this class via a Builder class. If you want to do it, you currently have to reimplement the whole class just for adding some fields because there is no non-ugly way to access the fields. My proposal here is to make them protected, so you can override the default build() method behaviour in your derived class and still have access to the properties.