Eliminate unnecessary query for current version
The version from query result is nondeterministic, it's not definitely same to the moment update executed because it may be altered right after that moment, it's not definitely same to the latest version in database because it may be altered right after the query executed. Given that, the current version is not so useful for troubleshooting in practice.

Signed-off-by: Yanming Zhou <zhouyanming@gmail.com>