Fix NullPointerException in MongoSequenceIncrementer
When using a MongoDB repository for the first time, the collections
used by MongoSequenceIncrementer don't yet exist, which results
in the call to findOneAndUpdate().getLong() to reutrn a null,
which then results in an NPE unboxing the Long into a long.

Adding upsert(true) creates the initial document with a
count of 0 and immediately increments it to 1.

Signed-off-by: blake_bauman <blake_bauman@apple.com>