Allow keys other than _id for primary key in MongoItemWriter
When moving data between different data store types, such as
between MongoDB and a SQL database, not all systems have a
"_id" as their pimary key.  Most often it has a different
name or can be multiple keys/columns used as a composite
key.  This change allows the app to specify an alternate
key or a set of keys which, together, can be used as the
primary key for a MongoDB insert/update/upsert.

Signed-off-by: blake_bauman <blake_bauman@apple.com>