Improve error message when using ResourcelessJobRepository with partitioned steps
Issue spring-projects/spring-batch#4732

The ResourcelessJobRepository cannot be used with partitioned steps as it
does not support execution context, which is required for partition metadata
sharing between manager and worker steps.

This commit adds an explicit check in SimpleStepExecutionSplitter to detect
usage of ResourcelessJobRepository and throw a clear, informative error
message instead of the confusing 'Cannot restart step from STARTING status'
message that was previously displayed.

Signed-off-by: khj68 <junthewise@gmail.com>