Move job parameter name from JobParameters to JobParameter
Before this commit, the name of a job parameter lived outside the
`JobParameter` class (used as a key of the internal map of `JobParameters`)
This had a few issues as explained in #5010.

This commit moves the name of a job parameter to the `JobParameter`
class. Moreover, this commit changes `JobParameter` and `JobParameters`
from being regular classes to records.

Resolves #5010