Remove job execution id from StepExecutionRequest
Step execution IDs are globally unique, and a step
execution always belongs to a job execution. This
commit removes the need to provide a job execution
id when sending step execution requests as this is
not used by clients (StepExecutionRequestHandler).

The job execution ID is still used in the correlation
ID of the message but is not part of the step execution
request anymore.