Update default configuration of Jackson 3 mapper in JacksonJsonObjectReader
The default value of DeserializationFeature.FAIL_ON_TRAILING_TOKENS has
changed from false to true in Jackson 3. This change requires users to
disable that feature to be able to use Spring Batch's JsonItemReader
correctly.

This commit disables that feature by default in JacksonJsonObjectReader
so that Spring Batch users do not have to do it manually.

Resolves #5047