Persist step execution status before calling listeners
Before this commit, the step execution status was not persisted
in the job repository before calling step execution listeners.
Therefore, listeners see the step's status as STARTED and not
COMPLETED or FAILED. This was not coherent with the contract of
`afterStep` which states that the method is called after the
execution of the step's processing logic (whether successful or
failed), where the status (both in memory and in the DB) should
not be STARTED at that point, but rather a non-running status.

This commit adds a job repository update to persist the step's
status before calling step execution listeners.

Resolves #4362