Improve batch components registration as listeners and streams
Before this commit, batch components were registered as listeners only
if they implement StepExecutionListener.

This commit enlarges the scope and registers them as listeners if they
implement StepListener (which is the super type of all listeners).

This commit also improves the step configuration by automatically
registering batch components as streams if applicable.

Resolves #5087